% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{---------------Multiply two numbers.----------------}
\usage{
multiply(val1, val2)
}
\arguments{
\item{val1}{A number}

\item{val2}{a number}
}
\description{
---------------Multiply two numbers.----------------
}
\examples{
multiply(2,3)
multiply(4,5)
}
